Fix clippy lints

+++ wow_world_base/src/inner/wrath/cache_mask.rs
@@ -51,7 +51,7 @@ impl TryFrom<u32> for CacheMask {
         match value {
             21 => Ok(Self::GlobalCache),
             234 => Ok(Self::PerCharacterCache),
-            v => Err(crate::errors::EnumError::new("CacheMask", v as u32),)
+            v => Err(crate::errors::EnumError::new("CacheMask", v),)
         }
     }
 }
