+++ wow_world_messages/src/world/shared/activate_taxi_reply_vanilla_tbc_wrath.rs
@@ -99,7 +99,7 @@ impl TryFrom<u32> for ActivateTaxiReply {
             10 => Ok(Self::PlayerMoving),
             11 => Ok(Self::SameNode),
             12 => Ok(Self::NotStanding),
-            v => Err(crate::errors::EnumError::new("ActivateTaxiReply", v as u32),)
+            v => Err(crate::errors::EnumError::new("ActivateTaxiReply", v),)
         }
     }
 }
