+++ wow_world_messages/src/world/shared/arena_team_command_error_tbc_wrath.rs
@@ -111,7 +111,7 @@ impl TryFrom<u32> for ArenaTeamCommandError {
             19 => Ok(Self::ArenaTeamIgnoringYouS),
             21 => Ok(Self::ArenaTeamTargetTooLowS),
             22 => Ok(Self::ArenaTeamTooManyMembersS),
-            v => Err(crate::errors::EnumError::new("ArenaTeamCommandError", v as u32),)
+            v => Err(crate::errors::EnumError::new("ArenaTeamCommandError", v),)
         }
     }
 }
