+++ wow_world_messages/src/world/shared/buyback_slot_vanilla_tbc_wrath.rs
@@ -99,7 +99,7 @@ impl TryFrom<u32> for BuybackSlot {
             79 => Ok(Self::Slot11),
             80 => Ok(Self::Slot12),
             81 => Ok(Self::Slot13),
-            v => Err(crate::errors::EnumError::new("BuybackSlot", v as u32),)
+            v => Err(crate::errors::EnumError::new("BuybackSlot", v),)
         }
     }
 }
