+++ wow_world_messages/src/world/shared/gm_ticket_response_vanilla_tbc_wrath.rs
@@ -69,7 +69,7 @@ impl TryFrom<u32> for GmTicketResponse {
             4 => Ok(Self::UpdateSuccess),
             5 => Ok(Self::UpdateError),
             9 => Ok(Self::TicketDeleted),
-            v => Err(crate::errors::EnumError::new("GmTicketResponse", v as u32),)
+            v => Err(crate::errors::EnumError::new("GmTicketResponse", v),)
         }
     }
 }
