+++ wow_world_messages/src/world/shared/new_item_chat_alert_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u32> for NewItemChatAlert {
         match value {
             0 => Ok(Self::DoNotShow),
             1 => Ok(Self::Show),
-            v => Err(crate::errors::EnumError::new("NewItemChatAlert", v as u32),)
+            v => Err(crate::errors::EnumError::new("NewItemChatAlert", v),)
         }
     }
 }
