+++ wow_world_messages/src/world/shared/new_item_source_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u32> for NewItemSource {
         match value {
             0 => Ok(Self::Looted),
             1 => Ok(Self::FromNpc),
-            v => Err(crate::errors::EnumError::new("NewItemSource", v as u32),)
+            v => Err(crate::errors::EnumError::new("NewItemSource", v),)
         }
     }
 }
