+++ wow_world_messages/src/world/shared/quest_completable_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u32> for QuestCompletable {
         match value {
             0 => Ok(Self::NotCompletable),
             3 => Ok(Self::Completeable),
-            v => Err(crate::errors::EnumError::new("QuestCompletable", v as u32),)
+            v => Err(crate::errors::EnumError::new("QuestCompletable", v),)
         }
     }
 }
