+++ wow_world_messages/src/world/shared/raid_group_error_vanilla_tbc_wrath.rs
@@ -44,7 +44,7 @@ impl TryFrom<u32> for RaidGroupError {
         match value {
             1 => Ok(Self::Required),
             2 => Ok(Self::Full),
-            v => Err(crate::errors::EnumError::new("RaidGroupError", v as u32),)
+            v => Err(crate::errors::EnumError::new("RaidGroupError", v),)
         }
     }
 }
