+++ wow_world_messages/src/world/shared/timer_type_vanilla_vanilla_tbc_wrath.rs
@@ -56,7 +56,7 @@ impl TryFrom<u32> for TimerType {
             1 => Ok(Self::Breath),
             2 => Ok(Self::FeignDeath),
             3 => Ok(Self::Environmental),
-            v => Err(crate::errors::EnumError::new("TimerType", v as u32),)
+            v => Err(crate::errors::EnumError::new("TimerType", v),)
         }
     }
 }
