+++ wow_world_messages/src/world/tbc/battleground_type.rs
@@ -79,7 +79,7 @@ impl TryFrom<u32> for BattlegroundType {
             6 => Ok(Self::Arena),
             7 => Ok(Self::EyeOfTheStorm),
             8 => Ok(Self::RuinsOfLordaeron),
-            v => Err(crate::errors::EnumError::new("BattlegroundType", v as u32),)
+            v => Err(crate::errors::EnumError::new("BattlegroundType", v),)
         }
     }
 }
