+++ wow_world_messages/src/world/vanilla/cinematic_sequence_id.rs
@@ -79,7 +79,7 @@ impl TryFrom<u32> for CinematicSequenceId {
             101 => Ok(Self::Gnome),
             121 => Ok(Self::Troll),
             141 => Ok(Self::Tauren),
-            v => Err(crate::errors::EnumError::new("CinematicSequenceId", v as u32),)
+            v => Err(crate::errors::EnumError::new("CinematicSequenceId", v),)
         }
     }
 }
