+++ wow_world_messages/src/world/vanilla/hit_info.rs
@@ -123,7 +123,7 @@ impl TryFrom<u32> for HitInfo {
             32768 => Ok(Self::Crushing),
             65536 => Ok(Self::NoAction),
             524288 => Ok(Self::SwingNoHitSound),
-            v => Err(crate::errors::EnumError::new("HitInfo", v as u32),)
+            v => Err(crate::errors::EnumError::new("HitInfo", v),)
         }
     }
 }
