+++ wow_world_messages/src/world/vanilla/mail_result.rs
@@ -101,7 +101,7 @@ impl TryFrom<u32> for MailResult {
             17 => Ok(Self::ErrMailAndChatSuspended),
             18 => Ok(Self::ErrTooManyAttachments),
             19 => Ok(Self::ErrMailAttachmentInvalid),
-            v => Err(crate::errors::EnumError::new("MailResult", v as u32),)
+            v => Err(crate::errors::EnumError::new("MailResult", v),)
         }
     }
 }
