+++ wow_world_messages/src/world/wrath/aura_type.rs
@@ -1667,7 +1667,7 @@ impl TryFrom<u32> for AuraType {
             314 => Ok(Self::PreventResurrection),
             315 => Ok(Self::UnderwaterWalking),
             316 => Ok(Self::PeriodicHaste),
-            v => Err(crate::errors::EnumError::new("AuraType", v as u32),)
+            v => Err(crate::errors::EnumError::new("AuraType", v),)
         }
     }
 }
