+++ wow_world_messages/src/world/wrath/emote.rs
@@ -909,7 +909,7 @@ impl TryFrom<u32> for Emote {
             474 => Ok(Self::StateReadySpellOmni),
             475 => Ok(Self::StateHoldJoust),
             476 => Ok(Self::OneshotCryJaina),
-            v => Err(crate::errors::EnumError::new("Emote", v as u32),)
+            v => Err(crate::errors::EnumError::new("Emote", v),)
         }
     }
 }
